fix(ui-ux): update latest transactions ui
